dev_log_archive_and_testfiles/requirements.txt
# Core Dependencies
anthropic>=0.39.0
streamlit>=1.40.0
streamlit-ace>=0.1.1
python-dotenv>=1.0.0

# Database & Vector Storage
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Scientific & Math Libraries
numpy>=1.24.0
scipy>=1.11.0
sympy>=1.12
mpmath>=1.3.0
torch>=2.0.0

# Quantum Computing (optional)
qiskit>=0.45.0
qutip>=4.7.0

# Code Quality & Security
black>=23.0.0
RestrictedPython>=6.0
sqlparse>=0.4.4
jsbeautifier>=1.14.0

# Networking & Web
requests>=2.31.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
aiofiles>=23.0.0

# Data Structures & Games
networkx>=3.1
chess>=1.10.0
pygame>=2.5.0

# Authentication & Security
passlib>=1.7.4

# Git Operations
pygit2>=1.13.0

# Time & Async
ntplib>=0.4.0
nest-asyncio>=1.5.8

# Data Validation
marshmallow>=3.20.0

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0

# Token Counting
tiktoken>=0.5.0

# YAML Processing
pyyaml>=6.0
